requirements.txt
asgiref==3.8.1
Django==4.2
python-decouple==3.8
pytz==2024.1
sqlparse==0.5.0
